tools/paper-processor/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paper-processor"
version = "1.0.0"
description = "Extract and process content from academic papers"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Deep Biblio Tools Team", email = "team@deepbiblio.tools"}
]
keywords = ["papers", "extraction", "pdf", "html", "academic", "processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "PyPDF2>=3.0",
    "pdfplumber>=0.9",
    "markdownify>=0.11",
    "rich>=13.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
summarization = [
    "literature-reviewer>=1.0",
]
advanced-pdf = [
    "pymupdf4llm>=0.0.12",
    "marker-pdf>=0.2.15",
]
all = [
    "literature-reviewer>=1.0",
    "pymupdf4llm>=0.0.12",
    "marker-pdf>=0.2.15",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
paper-processor = "paper_processor.cli:main"

[project.urls]
Homepage = "https://github.com/deep-biblio-tools/deep-biblio-tools"
Documentation = "https://deep-biblio-tools.readthedocs.io/tools/paper-processor"
"Bug Tracker" = "https://github.com/deep-biblio-tools/deep-biblio-tools/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_processor"]
